+++ src/oihana/routes/enums/NoBodyResponse.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace oihana\routes\enums;
+
+use oihana\reflect\traits\ConstantsTrait;
+
+/**
+ * Special response indicators
+ */
+class NoBodyResponse
+{
+    use ConstantsTrait ;
+
+    const string DEFAULT = 'noBodyResponse'    ;
+    const string CLAZZ   =  ':noBodyResponse'  ;
+}