+++ src/oihana/routes/http/DeleteRoute.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace oihana\routes\http;
+
+use oihana\enums\http\HttpMethod;
+
+class DeleteRoute extends MethodRoute
+{
+    const string INTERNAL_METHOD = HttpMethod::delete ;
+}