+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - Adds the oihana\controllers package (beta)
+- Adds the oihana\routes package (beta)
 
 - Adds oihana\date\TimeInterval
 - Adds oihana\date\traits\DateTrait